SC30-007 add missing calls to Retysp
In both marking and proof we were tested a type with Has_Fixed_Point_Type,
which relies on a Retysp, but then we queries the Small_Value directly from
the Etype, which was inconsistent.

* gnat2why/spark/spark_definition.adb
(Mark): Call Small_Value on the Retysp, not on the Etype.

* gnat2why/spark/spark_definition.
(Transform_Expr): Likewise.

* testsuite/gnatprove/tests/SC30-007__private_fixed_type/
Now is silently accepted, as it should.

Change-Id: I54095f9c5690b350f87e546a2665a8e016d06a34