T117-040 suppress warning on null assignments via subprogram calls
When a parameter is assigned via a subprogram call and an explicit Depends
contract says that this assignment has a null dependency, then we suppress
an "unused assignment" warning.

This warning was particularly annoying on calls to instances of
Ada.Unchecked_Deallocation, but appears to be useful in general.

* gnat2why/flow/flow-analysis.adb
(Null_Dependency_Assignment): New routine for detecting when to suppress
 the annoying warning.

(Find_Ineffective_Statements): Suppress warnings using the newly added
 routine.

* testsuite/gnatprove/tests/QA23-042__error_pointers/test.out
* testsuite/gnatprove/tests/RA12-032__pointer_free/test.out
* testsuite/gnatprove/tests/S117-008__pointer/test.out
* testsuite/gnatprove/tests/T113-042__storage_size/test.out
Update; annoying warnings with instances of Ada.Unchecked_Deallocation are
gone.

* testsuite/gnatprove/tests/T117-040__flow_null_depends/test.out
Update; annoying warning is now suppressed.

Change-Id: I4bd68ebd653c7d565640c852a629c66ce0e2dad1