SC02-022 fix code for pragma Annotate on subp body
The previous code that dealt with pragma Annotate on
N_Subprogram_Body nodes tried to work around a front-end bug:
Sloc_Range would return incorrect results when called with a
N_Subprogram_Body node. However, the work-around was incorrect, in
some cases, and a nonsensical/wrong range computed.

The bug in the front-end was fixed on T121-001, so we can remove the
workaround.

* spark_annotate.adb
(Insert_Annotate_Range): we can now call Sloc_Range on the
  N_Subprogram_Body directly.
(Insert_Annotate_Range): remove now-unused variant

Change-Id: I9028cf0fcdb807892e44824c82f251347a7dc235