S611-030 add test showing flow visibility violation
GNATprove crashes with an assertion failure when flow visiblity tries to
access a private child procedure that overloads a public parent
procedure of the same name. This reproducer illustrates the crash.

Flow should respect information hiding and not peek into private parts
that are not visible. Flow should pretend that the call is to the
inherited routine.

* testsuite/gnatprove/tests/S611-030__flow_visibility_private_child_...
Add test; disable because it currently crashes.

Change-Id: Iace69df53e36d77dd34055ffb46f597bdb5d64f9