SA22-034 change set-based Up_Project procedures to functions
Historically, we were always up-projecting sets, so it made sense to have a
procedure that takes and returns sets. After fixes on this TN some of our
up-projections are only called with singleton sets, so it is simpler to
change those procedures to functions. Semantics unaffected.

* gnat2why/flow/flow_generated_globals-phase_2.adb
(Up_Project): Variant for Name_Sets is now a function; callers adapted.

* gnat2why/flow/flow_refinement.adb
(Up_Project): Variant for Flow_Id_Sets is now a function; callers adapted.

Change-Id: I005080a6a9c7e0fd6ec964bf686587d4f64b028a