S611-030 relax visibility assertion for hidden dispatching operations
Before this patch, we used to check that the caller could see the called
subprogram in the flow visibility graph, when constructing the control
flow graph. This is a valid check to perform, but it does not handle the
case of calling dispatching operations declared in a private part - thus
breaking the assertion.

This patch relaxes the assertion to cater to calls to hidden dispatching
operations.

* src/flow/flow-control_flow_graph.adb
(Do_Call_Statement): Handle hidden dispatching operations

* testsuite/gnatprove/tests/S611-030__flow_visibility_privat.../test.out
Enable test; add test output.

Change-Id: Ic470e88cec11b9430f85ca16d8544a5d65df680c