T210-005 move test to large testsuite
This test runs for ~20seconds even with --cache and slows down the typical
runs with --cache by almost as much. Also, it primarily guards against
regressions related to the Depends contracts in the formal containers
library, so it is unlikely to be very useful in everyday hacking.

* testsuite/gnatprove/tests/T210-005__conts/test.opt
Move to large testsuite.

Change-Id: I10ed355276dd3a68e504d7dce53efad2b4d153e7