T212-053 Skip past subtypes for generic instantiations
Retysp of a type should not be the subtype introduced by the frontend
for the instantiation of a generic formal type. Use the actual instead
when possible. Otherwise, the representative pointer type for example is
not correctly computed leading to conversion errors.

* src/spark/spark_util-types.adb
(Retysp): Replace return statements by gotos to the end of the function,
 in order to get past the generic actual subtype if needed.

* testsuite/gnatprove/tests/T212-053__generic_pointer
Add test.

Change-Id: I0c446d98d4579d9b51ad69ae17f69f2b4b31203a