T207-035 add a similar test but with type conversion
A similar problem applies both to type qualification (which is the original
subject of this TN and we already have a test for that) and to type
conversion (for which a test is added here): referencing a type in qualified
expression or type conversion does not pull its constraints as data
dependencies.

* testsuite/gnatprove/tests/T207-035__flow_constrained_depends/
Add a similar test, but with type conversion instead of qualification;
the Depends contract is correct and flow checks are wrong.

Change-Id: I1eb625bc13b018b13af1e682725172a6832210f2