T207-035 avoid dependency from type conversion
When traversing the AST and seeing an N_Type_Conversion node like
"My_Type (Expr)" we were processing the subtype indication (here "My_Part")
and if it happened to be a constrained type, we were pulling dependencies
from its constraints. This was wrong and inconsistent wrt handling of
N_Type_Conversion in Untangle_Record_Assignment.

* src/flow/flow_utility.adb
(Get_Variables_Internal): For N_Type_Conversion that applies to a non-record
 type only process its Expression field and skip all the other fields.

* testsuite/gnatprove/tests/T207-035__flow_constrained_depends/test.out
Update; spurious checks are gone.

Change-Id: Icbc7e7982aaea1c25e621bf7f4c6d559d2aa5a1f