T207-035 avoid dependency from unchecked conversion
When traversing the AST and seeing an N_Unchecked_Type_Conversion we were
processing its target subtype and if it happened to be a constrained, we
were pulling dependencies from its constraints. This was wrong and
inconsistent wrt handling in Untangle_Record_Assignment.

* src/flow/flow_utility.adb
(Get_Variables_Internal): For N_Unchecked_Type_Conversion only process its
 Expression field and skip all the other fields.

* testsuite/gnatprove/tests/T207-035__flow_constrained_unchecked/test.out
Update; spurious checks are gone.

Change-Id: I16a91765648f9458dacf066f439ba66718870f14