T207-035 refactor sanity checks for variable inputs
The driving idea for this refactoring is to add a precondition to
Detect_Variable_Inputs so that we call it only on subexpressions and not on
entire type declarations (or their parts). Also, we now only check
declarations that come from source.

This refactoring is certainly does not preserving semantics, but we
already have a quite comprehensive test coverage for variable inputs
and there are no diffs (other that messages being located precisely at the
offending expressions), which demonstrates that this refactoring works.

* src/flow/flow-analysis-sanity.adb
(Check_Expressions): Refactor.
(Detect_Variable_Inputs): Add precondition.

* src/flow/flow_utility.ads
* src/flow/flow_utility.adb
(Get_Variables, Get_Variables_Internal): Remove variants for List_Id, as
 they are no longer used.

* testsuite/gnatprove/tests/P517-019__flow_types/test.out
* testsuite/gnatprove/tests/Q412-030__type_invariant/test.out
* testsuite/gnatprove/tests/Q925-025__flow_type_bound/test.out
* testsuite/gnatprove/tests/R212-038__flow_return_varinput/test.out
* testsuite/gnatprove/tests/S527-027__flow_variant_part/test.out
* testsuite/gnatprove/tests/S529-037__inlining/test.out
* testsuite/gnatprove/tests/S530-006__internal_variable_input/test.out
Update locations.

Change-Id: If03505f713201d90a034b8c5a20dca9886018f4f