T207-035 add test for wrong dependency on allocator
When an object is allocated using the first variant from Ada RM 4.8(2/3):

   allocator ::=
      new [subpool_specification] subtype_indication    <<< this test
    | new [subpool_specification] qualified_expression

and the subtype_indication refers to a constrained scalar type, then the
bounds of this type are wrongly pulled as data dependencies. This is similar
to other tests for this TN, where such a dependency is pulled from a type
conversion, qualified expression or quantified expression.

* testsuite/gnatprove/tests/T207-035__flow_constrained_allocator/
Add; the Depends contract is right, while flow checks are wrong.

Change-Id: I27917a7674e28f09b80420965714241a4d91d164