T227-051 fix getting variables from slices with subtype indication
For a slice range is given as a subtype indication, e.g.:

   Arr (My_Integer range Low .. High)

we get variables only from the range bounds and not from the bounds of the
subtype (i.e. from Low and High but not from My_Integer).

* src/flow/flow_utility.adb
(Get_Variables_Internal): Explicitly traverse range bounds and the prefix,
 while skipping other nested nodes.

(Untangle_Assignment_Target): Explicitly traverse bounds of the range and
 not the range expression itself (which might be a subtype indication).

* testsuite/gnatprove/tests/T227-051__subtype_slice/test.out
Spurious checks are gone.

Change-Id: Ia18948a5a4ebc5d172990c62147428ca965f3713