T227-051 checks for subtype indication in aggregates and slices
A subtype indication can be used in a slice and in an array aggregate
in place of a range. In this case, we must emit a check to make sure
that the bounds are in the provided subtype.

* src/spark/spark_util-types.adb
(Is_Null_Range): May be called on a subtype indication. Use Type_*_Bound
 to retrieve the bounds.

* src/why/gnat2why-expr.adb
(Transform_Aggregate): When translating an array aggregate, possibly emit
 checks for range constraints in subtype indications.
(Transform_Slice): If the range of the slice is a subtype indication, emit
 range checks.

* tests/M830-004__checks/test.out
Additional range checks on subtype indications.

* tests/T227-051__subtype_slice
Add slight variant reported on T227-050 and enable proof.

Change-Id: Ibff700bcac3c67bda1aa4031a581b7993416371a