mark allocators in top-down style
Marking of allocators as a sequence of conditions was slightly hard to
follow. Now it is done in a top-down style; in particular, we clearly
separate marking of initialized and uninitialized allocators.

(no-tn-check)

* src/spark/spark_definition.adb
(Mark): Rewrite marking of N_Allocator; remove repeated checks.

* testsuite/gnatprove/tests/O416-002__dynamic_predicate_not_in_spark/...
Update; the reason for marking failure is different now.

Change-Id: Ie486095dbe40901ac5c7b8118354261c25489fcb