T207-035 fix dependencies from scalar attributes
When getting objects referenced in an attribute reference of the form
"prefix'attribute (expressions)" and the prefix denotes a scalar type, we
already special-cased processing of 'First/Last/Length/Range. For other
attributes we should ignore their prefix and not pull dependencies from the
scalar bounds.

* testsuite/gnatprove/tests/T207-035__flow_constrained_depends/test.out
Add; spurious checks are gone.

Change-Id: I8e63caba975f8168623f96db9fe90076d9b9dc5b