T318-005 adapt test after a frontend fix
A code that looked like this:

   package P is
      type T is limited private;
      type A is array (...) of T
      with Pack, Volatile, Alignment => ...;  --  representation aspects
   private
      type T is record ... end record;
   end;

is now rightly rejected, after GNAT correctly implements Ada RM 13.1(9.2/5).
This rule says that representation aspects shall not be specified before the
type is completely defined; in particular, they for an array type it shall
not be specified before the full declaration of the array component.

* testsuite/gnatprove/tests/QA04-034__volatile/hw-dbc-trbs.ads
Adapt test by making the array component type non-private.

Change-Id: I3d952c0e969d71e4fa6f0fee8c83d0370391eb2c