T326-001 fix proper scoping of names for clone substitutions
We used the string "I1.t" (for example) as an identifier name for
the left-hand-side of clone substitutions, which required
postprocessing in why3 to extract module name and local name. We now
properly construct a name with module "I1" and local copmonent "t".
To avoid including such modules in the set of dependencies, we skip
them during the traversal of the Why tree which looks for
dependencies.

* why-gen-arrays.adb
(Prepare_Index_Substitutions): Use properly scoped names for clone
  substitutions
* why-inter.adb
(Compute_Module_Set): don't look at left-hand-side of clone
substitutions

Change-Id: I2249ad08870fec41860adee356ac8de7cce72b72