Update Loop_Entry attribute rules to match new Old attribute rules
AI12-0217 made changes in the rules for 'Old; make analogous changes
for 'Loop_Entry. In particular, where the rules used to require that
the attribute prefix must "statically denote" an object in some cases,
those rules have been relaxed to say that the prefix must "statically
name" an object instead. The significance of this is that it is now
ok, at least in many cases, to name a subcomponent of a top-level
object in the attribute prefix, as opposed to only being able to name
the top-level object.

* docs/lrms/source/statments.rst: Replace a couple of uses of "statically
  denote" with "statically name" in the legality rules for the Loop_Entry
  attribute. Also clean up slightly the phrasing of some non-normative text.

Change-Id: I92178dc888033b084b10b39d90a0eb623c3102ef
TN: T326-043