T211-001 fix generation of JSON file for menus
The Makefile was previously calling `why3 --list-transforms`, but
this requires the why3 binary to be installed in
<prefix>/libexec/spark/bin. But we don't want the binary to be
copied there, and the path doesn't necessarily exist (because it is
created during make install, which is called later).

We fix this by using `gnatwhy3 --list-transforms`, which also works
when not installed in the proper location.

Change-Id: I313a8494397f99ae4a25ec744ec102437ff2dba4