T127-031 minimal fix to not process subprogram aliases
Rewriting of AST done at the very beginning of the backend processing didn't
traverse type invariant expressions, so marking was pulling for translation
subprogram aliases called in those expression.

Note: after this fix most of the calls to Ultimate_Alias in the backend can
be eliminated, but for clarity and sanity it seems better to remove them in
follow-up commits.

* src/spark/spark_definition.adb
(Mark_Call): Sanity-check that we no longer process subprogram aliases.

* src/spark/spark_rewrite.adb
(Rewrite_Node): Traverse type invariant expressions; an extra guard with
Comes_From_Source it to ignore repeated type declarations for derived types
that are introduced by the frontend (it is not essential for this fix, but
it also makes the traversal of type invariant closer to what it should be).

* src/spark/spark_register.adb
(Process_Node): Mirror change in Rewrite_Node (it is a known issue that
those two routines should be refactored to avoid duplication and slight
differences).

* src/spark/spark_util-subprograms.ads
(Get_Expr_From_Check_Only_Proc): Accept partial invariant procedures.

* testsuite/gnatprove/tests/R514-018__type_inv/test.out
Update output; new check is legitimate.

Change-Id: I1687c2dab0e2bc8ef1ff1708d289f294e0ea2957