T127-031 remove unnecessary Get_Iterable_Type_Primitive wrapper
The SPARK_Util.Types wrapper for Sem_Util.Get_Iterable_Type_Primitive was
added to call Ultimate_Alias on the original routine's result. It could be
necessary when this routine was added, though now is it no longer
needed (even before any fixes submitted on this TN).

The iterable type primitives can only appear from an explicit aspect and can
only point to user-defined subprograms. This aspect is first translated into
pragma, and pragma is attached to the representation item chain of the type.

The mapping from type to the First/Next/Has_Element subprograms do not
change when a derived type is created; instead, the derived type simply
inherits the mapping, which still points to a user-defined subprogram for
the parent type (this happens in Inherit_Rep_Item_Chain).

Finally, this wrapper was enforcing a trivial precondition, but it was
unnecessarily restrictive (though in the backend we don't use the Last and
Previous primitives at the moment).

* src/spark/spark_util-types.ads
* src/spark/spark_util-types.adb
(Get_Iterable_Type_Primitive): Remove wrapper.

* src/why/spark_atree-entities.ads
* src/why/spark_atree-entities.adb
(Get_Iterable_Type_Primitive): Add a simple renaming with a stronger
contract.

* src/spark/spark_annotate.adb
* src/why/gnat2why-expr-loops.adb
* src/why/gnat2why-expr.adb
Adapt callers.

Change-Id: I82588eac5a66dadf72a9a150e86669c01dbc4df2