T127-031 don't expect aliases in the Iterable aspect expression
Names of functions in the Iterable aspect expression can only refer to
user-defined functions and never to aliases created by the frontend for
derived types. This is now checked in a contract of Mark_Subprogram_Entity,
which ultimately processes those functions.)

* src/spark/spark_definition.adb
(Mark_Iterable_Aspect): Remove misleading comment and unnecessary code.

Change-Id: I8e3c1669603aac5dc30fa309b7b950c16d7ee7d0