Only queue type-related subprograms for marking if type is in SPARK
We were queuing predicate functions and direct primitive operations for
marking even if the type itself was decided to be not in SPARK. Now we only
queue them if the type is in SPARK, which follows the top-down marking
approach.

no-tn-check

* src/spark/spark_definition.adb
(Mark_Entity): Guard code for queuing type-related subprograms with a check
of Violation_Detected.

Change-Id: I725575c72a70298d8162875e01ad79f9c423eaa8