T407-036 refactor handling of procedure calls
The handling of procedure calls is separated in two parts. The computation
of the call parameters, done in Compute_Call_Args, and the introduction
of a context in Insert_Ref_Context, necessary to handle complex mutable
parameters (either involving a conversion, or referring to a part of a
data-structure). The two subprograms were heavily interdependant, and
handling of each concern (aliasing, initialization...) was scatered in
several places.
Refactor to do all the computation in Compute_Call_Args, and only do the
reconstruction in Insert_Ref_Context. Gather handling of each concern in
one place as much as possible.

* src/why/why-atree-modules.adb
(Insert_Type_Symbols): Clean-up, use EW_Abstract like everywhere else.

* src/why/why-gen-expr.ads
(Needs_Slide): Moved from Insert_Array_Conversion to the top-level so that
 it can be reused.

* src/why/gnat2why-expr.adb
(Ref_Type): Type for mappings from an identifier to an expression.
(Ref_Context): Vector of mappings.
(Get_Item_From_Expr): Split a Why expression into parts that will be used
 to call a subprogram. References are introduced for mutable parts. Take
 as input an item representing the expected form of the formal parameter.
 Its variable  parts are used for the names of the new references introduced
 for the variable parts. Mappings for these variables are stored in a
 ref_context.
(Get_Item_From_Var): Try to reuse parts of the references of the actual for
 the formal. If the types do not match, fall back to Get_Item_From_Expr.
(Compute_Store): Compute the sequence of statements necessary to store back
 local identifiers of an item inside an actual. Also introduce
 checks when necessary.
(Compute_Call_Args): Create an item from the formal entity to serve as a
 pattern providing fresh names that can be used for new references. Use
 Get_Item_From_Expr and Get_Item_From_Var to collect the appropriate parts
 of actual parameters to use for the call. Use Compute_Store to compute the
 matching store operation. Return the context, the store and the arguments
 to its caller.
(Insert_Ref_Context): Take as input the Why3 call, the context and the
 store and glue everything together to model the Ada call.

* tests/K518-029__call_actuals/test.out
* tests/S117-008__pointer/test.out
Counterexample change and a new check is proved due to a change in the
order in which we check parameters for absence of RTE.

* tests/K607-022__unconstrained_array/test.out
* tests/T212-013__relaxed_init/test.out
* tests/T212-013__relaxed_init_loops/test.out
Some trivialy true length checks are no longer emitted as we recognize
that they are not needed.

* tests/KB25-003__slices/test.out
* tests/M611-052__discr/test.out
* tests/OC01-011__justify_aliasing/test.out
* tests/R427-016__check_on_param_call/test.out
* tests/R902-002__size_and_length/test.out
* tests/S219-014__nested_update/test.out
* tests/adacore_u/test.out
Counterexamples change due to a change in the order in which we check
parameters for absence of RTE.

* tests/T407-036__out_param
New test.

Change-Id: I85624264dd8ed93992d616e91d43fa4193d60ec4