SPARK RM wording for Relaxed_Initialization aspect
The Relaxed_Initialization aspect and the related Initialized attribute
are defined to allow safe usage of partially initialized composite objects.
This is just an initial version - the precise syntax rules for
Relaxed_Initialization aspect specifications for subprograms are missing
and there are some formatting problems (e.g., indentation of code
snippets). This is just to get initial feedback.

* docs/lrms/source/subprogram.rst: Provide rules for the
  Relaxed_Initialization aspect and for the Initialized attribute.

Change-Id: I6e51c572e0094f9d48a06d3c04268469525d1339
TN: T212-014