T412-002 refactor checking of annotate pragmas to be part of marking
* src/spark/spark_definition-annotate.adb
Renames src/spark/spark_annotate.adb.
(Check_Inline_Annotation): Only look for an expression function body if
 the function has no contracts. Emit a violation and not a syntax
 error if the body of the expression function is not in SPARK.

* src/spark/spark_definition-violations.ads
SPARK_Definition.SPARK_Violations is now a private child unit so it can be
called form Check_Inline_Annotation.

* src/spark/spark_definition.ads
(SPARK_Violations): Factored out.
(SPARK_Pragma_Is): Declared in the private part so it can be called form
 Violations.
(Emit_Messages,
 Current_SPARK_Pragma,
 Current_Delayed_Aspect_Type,
 Current_Incomplete_Type): Moved to Violations as they are used for error
 reporting.

* tests/RA14-002__effect_on_exit_path/test.out
Some checks were now unproved as the postcondition is taken instead of the
definition of an inlined for proof function if both are present. Remove the
postcondition of the inline annotation, it was useless.

* tests/T412-002__inline/
New test to exercize error messages with inline_for_proof.

Change-Id: Ib94ffe803d8b4f44da05147249f789cc150e4822