T414-005 fix crash in default initialization
The SPARK status of a record type is decided based on the SPARK status of
those of its components which are "visible in SPARK". We must rely on this
same "visible in SPARK" criteria when deciding default initialization, as
otherwise we will process components that might not be in SPARK.

* src/flow/flow_utility-initialization.adb
(Default_Initialization): Assume components that are not visible in SPARK
to be not default-initialized.

* testsuite/gnatprove/tests/T414-005__flow_default_init/
Crash is gone; enable. There is no output, because there is no SPARK_Mode.

Change-Id: Id2142f0cbb506ba51db481ba008ec29281943806