T227-053 move checks for value of uninitialized allocators
They are now located on the allocator instead of the subtype. This is
to better match what is done for execution and to avoid calling flow
analysis on node which are not expressions in generation of
explanations.

* src/why/gnat2why-expr.adb
(Transform_Expr): Use the allocator expression for the Ada node of the
 allocated value in uninitialized allocators.

* src/why/spark_atree.adb
(Do_Check_On_Scalar_Conversion, Get_Range_Check_Info): Expect the
 allocator for the Ada node of the allocated value in uninitialized
 allocators.

* src/flow/flow_error_messages.adb
(Get_Variables_From_Node): Remove special-casing for uninitialized
allocators.

* tests/RB14-019__subtype_checks/test.out
* tests/S305-038__allocators/test.out
* tests/T207-035__flow_constrained_allocator/test.out
* tests/T302-040__allocator_base/test.out
Checks on the allocated value of uninitialized allocators are now located
on the allocator.

Change-Id: I3a1ea344369c4eca4618f8abe8eff6de1689d4c6