T421-039 allow variable input in declarations of declare expressions
We have a blanket rule that forbids variable inputs in certain declarations.
This rule can be safely relaxed for declarations inside a declare
expression, because there is no risk that the variable will be modified
during the lifetime of such declarations.

This change is both to make the language as permissive as possible and to
relieve the flow analysis from having to detect variable inputs in every
expression (which would be a bit problematic to implement).

* docs/lrm/source/names-and-expressions.rst
(Expressions): Adapt rule about variable inputs to declare expression.

Change-Id: I6cb7665716c1d1ab4a1bc918e5251478eb661fe0