T429-023 add index check for choices of array delta aggregate
For delta aggregates and the expression of a 'Update attribute, index
checks are introduced for all indexes which are assigned as part of the
update. We should mirror these checks in gnat2why.
Also change the computation of bounds of ranges from choices to use their
Etype instead of the index type of the array as these bounds can be out
of range of the index subtype in case of empty ranges.
Also replace most mentions of 'Update by delta aggregates in preparation
for their deprecation.

* src/why/gnat2why-expr.adb
(Complete_Translation): Add index checks for array delta aggregates.
 Checks are located on the choice for 1 dimension arrays, and on actual
 indexes for multidimensional associations.
(Traverse_Value_At_Index): Use the bounds' Etype instead of the index type
 for ranges as bounds may be out of the index's subtype when the range
 is empty.

* src/why/spark_atree.adb
(Get_Range_Check_Info): We do not need to query the actual subtype for
 the choices of an array delta aggregate. This was ad hoc and only worked
 on parameter entities. Instead, we now separately insert index checks for
 these choices in Transform_Array_Aggregate.
 For bounds of ranges occuring in choices of delta aggregate, we use the
 bounds' Etype instead of the index type as they may be out of the index's
 subtype when the range is empty.

* tests/LC17-035__update_attribute/test.out
* tests/LC17-035__update_attribute_multidim/test.out
* tests/M930-005__flow_tick_update/test.out
* tests/M930-005__flow_update/test.out
* tests/MA22-017__old_in_update/test.out
* tests/MB12-007__unbound_symbol/test.out
* tests/N217-028__update_checks/test.out
* tests/N228-005__update_unconstr_array/test.out
* tests/N602-053__update/test.out
* tests/N708-013__flow_untangle_untangle/test.out
* tests/NC14-005__update/test.out
* tests/O512-022__vscomp2014_partition_rte/test.out
* tests/O522-028__loop_entry_update/test.out
* tests/RA14-002__effect_on_exit_path/test.out
* tests/RM_Examples/test.out
* tests/S219-014__nested_update/test.out
* tests/S724-006__string_update/test.out
* tests/T422-013__delta_on_arrays/test.out
* tests/TU__formal_parameters_illegal/test.out
* tests/riposte__simple_arrays/test.out
New index checks. They are either proved or duplicates of unproved range
checks.

* tests/T423-038__delta
New test.

Change-Id: Iefe6d9f05f3c1a4107303f3bf7ea542c8c0743d1