T430-026 Minor improvement in theory for fixed-points
The theory Fixed_Point_Shared for fixed-points was using Int_Division
which is mapped to ComputerDivision in Why3 theory, and ultimately to
an axiomatized definition for provers instead of the default div in
SMT-LIB2. As this symbol is only used for division of non-negative values,
it is equivalent to use EuclideanDivision, which is mapped to div in
SMT-LIB2 and thus leads to more checks proved (mostly in the private
testsuite).

Some regressions have been witnessed with CVC4, leading to changes in
the prover that proves some checks, as well as regressions in
counterexamples. The underlying issue is some fragility in how CVC4 proves
such VCs, which has been reported in CVC4 Issue Tracker under #4420

* share/spark/theories/ada__model.mlw
(Fixed_Point_Shared): Use EuclideanDivision.

* testsuite/gnatprove/tests/KB28-016__over/test.out
* testsuite/gnatprove/tests/M823-017__fixed_points/test.out
* testsuite/gnatprove/tests/O318-021__modular/test.out
* testsuite/gnatprove/tests/O429-042__stdlib_tasking/test.out
* testsuite/gnatprove/tests/O429-042__stdlib_tasking/test.py
* testsuite/gnatprove/tests/R220-054__fixed_point_small/test.out
* testsuite/gnatprove/tests/R722-002__manual/test.out
* testsuite/gnatprove/tests/R722-002__manual/test.py
* testsuite/gnatprove/tests/thumper/test.out
Changes in prover.

* testsuite/gnatprove/tests/O617-006__fxp_div_int_res/test.out
* testsuite/gnatprove/tests/RB12-066__fixed_point_div_mod/test.out
Changes in counterexample.

* testsuite/gnatprove/tests/O813-018__fixed_points/test.out
One more check proved. Changes in prover.

* testsuite/gnatprove/tests/RB12-066__fixed_point_int_mult/test.out
Change in counterexample. Additional checks duplicate existing range checks
at the same location, due to the structure of the Why3 code. As CVC4 does
not prove the original check here, it gets proved by Z3 while CVC4 proves
the duplicate one, leading to 2 checks instead of one.

Change-Id: Ib78918c282752958f0caf90f4412dd2caf0a02cc