T504-035 mark predicate in runtime units and update test outputs
After a recent change in the library of big numbers, we now have a
predicate on a subtype of big numbers in a runtime unit. We should mark
the associated predicate function. Also update test outputs.

* src/spark/spark_definition.adb
(Mark_Subprogram_Body): Mark the body of predicate functions even if they
 are located in library units as we will use them to get the predicate
 expression.

* tests/SB22-010__big_integers_support/test.out
* tests/T318-014__big_reals/test.out
New predicate checks.

Change-Id: I10550a6ac71865251c120c918a808fa830b7eb1e