P301-005 use declaration to find parent of derived type
As the Etype of a derived type may not be the subtype it is derived from,
use the declaration node to find the proper parent type of derived types.

* src/why/gnat2why-expr.adb
(Compute_Dynamic_Predicate): Use the type declaration to find the parent
 of a derived type.

* tests/P301-005__pred_rec_decl/test.out
Predicate check on default is now duplicated. This already happens on
several tests, is relatively harmless and hard to fix.

* tests/T505-002__predicate
Enable test.

Change-Id: I41000c423e4a2c6971e092526b730ba54970c4b1