T507-039 modify the test to avoid spurious check marker
The marker was positioned to avoid an unnecessary range check on an
initialized allocator with a qualified expression of the designated subtype.
Unfortunately, on the same line a range check was introduced by the
frontendfor the qualification. Modify the check so that no range check
is produced on the qualified expression either.

Change-Id: Ieb1f00db88d74d4b5ef098bc400a1292a07bc5aa