T504-004 refactor code for loop variants
The main change is to push most handling of loop variants into why3
code. To this end, the intermediate AST now contains nodes for loop
variants, which contain the loop variant expression as a term, and
the correct operator to compare new and old values. The actual
assertion for checking of loop variants is now expanded in the
translation to Ptree in Why3.

Another change is that, in the intermediate AST, loops are now
infinite loops (without exit condition) that permit loop invariants
and variants anywhere in the loop (together).

* xgen/xtree_sinfo.adb
  changes to the tree: new nodes for Variant (corresponds to an
  Increases/Decreases item in Ada) and Variants (corresponds to a
  pragma Loop_Variant in Ada). Loop node doesn't have a condition
  anymore, but gains a list of Variants, and Code_Before/Code_After
  nodes.
* why-sinfo.ads
  comments for new items.
* gnat2why-expr-loops.adb
(Transform_Loop_Variant): simplified to just translate the loop
  variant term to the intermediate Variant node.
(Check_Loop_Variant): new function to compute a program that checks
  for AoRTE for a loop variant.
(Wrap_Loop): adapt to new Loop node in intermediate tree
* testsuite:
  Checks for Loop_Variants are now placed on the Increases/Decreases
  instead of the pragma. Some minor changes in provability and CE.

Also fixes T518-001.

Change-Id: Idf44a69f1c9b97cea0d38229191fb9ca0ef26be6