T514-020 use the entity Id to mark a program specification
Change marking of subprogram specifications to use the entity_id instead
of the subprogram specification node. This makes it possible to reuse the
code for marking subprogram types. The only change is on the location of
the violation emitted for entry families.

* src/spark/spark_definition.adb
(Mark_Function_Specification, Mark_Subprogram_Specification): Expect a
 subprogram entity id instead of its specification node.

* tests/OA23-003__entry_family/test.out
* tests/P316-015__entry_family_call/test.out
Slight changes in locations of violations related to entry families.

Change-Id: If8700929f655a9a5e1a3e87e27a7ff2e8d7d6aee