T520-030 Aliased access argument
Add minimal reproducer copied from customer's support request,
and a further small testsuite for other cases involving
potentially aliasing access types and globals.
Fixed the implementation of Flow.Analysis.Antialiasing; it
previously assumed that all formal parameters of mode "In"
were not subject to aliasing but clearly (with access types)
this is no longer the case.
The pre-existing anti-aliasing code had not been implemented
in a way that was traceable to the SPARK RM section 6.4.2;
when the code was refactored to improve tracing as part of the
bugfix more small "holes" in the implementation were detected
with respect to access types (a new ticket has been raised).
Also fixed a bug (which could cause a precondition violation on
calling Is_Constant_After_Elaboration) which had been added by
me in a previous ticket (!)

* testsuite/gnatprove/tests/T520-030__flow_aliased_access/*
  Minimal reproducer supplied by the customer, further test
  cases developed for access types, plus test infrastructure.
* src/flow/flow-analysis-antialiasing.ads
  Improve description of Check_Procedure_Call
* src/flow/flow-analysis-antialiasing.adb
  (Is_Immutable) Explicit implementation of SPARK RM 6.4.2(2)
  (Is_By_Copy_Not_Access) Explicit implementation of common
  text in the second bullets of 6.4.2(3) and (4).
  (Aliasing) Fix by specifically implementing each clause from
  SPARK RM 6.4.2(3) and (4). Also fix a bug associated with
  implemeting 6.4.2(1).
  (Check_Procedure_Call) Numerous fixes associated with the
  necessity to check 'mode In' formal parameters, correcting
  the conditions under which we compare parameters and globals
  and improving debugging.

Change-Id: I592d9bd4bbf3d9cb9ced375f92fefcc684c95f59