Q502-001 Cleanup checking for SPARK violations
The marking code already checks at the start of Mark if an expression with
has a type allowed in SPARK. No need to recheck that at the beginning of
special handling for various node kinds.

The unreachable code was pointed at by the automatic coverage results.

* src/spark/spark_definition.adb
(Mark): Remove unreachable code.

* testsuite/gnatprove/tests/Q502-001__target_name/illegal_type.ads
* testsuite/gnatprove/tests/Q502-001__target_name/test.gpr
* testsuite/gnatprove/tests/Q502-001__target_name/test.out
* testsuite/gnatprove/tests/Q502-001__target_name/test_target_illegal.adb
Update test to have a case of type not in SPARK with @ construct.

Change-Id: Ib0cfbc19cc9df4f43ac1d87fd54551beb058d1f1