T525-030 fix crash in instances of predefined generic subprograms
In global generation we have a heuristic for instances of predefined generic
subprograms whose body has SPARK_Mode => Off. Those instances are callers
and the heuristic pulls actual subprograms of the instance as their callees.

This heuristic has an iteration to do deal with nested generics, e.g.:

  generic
    with function "=" ...          --  callee
  package Ada.Containers.Formal_Vectors is
     ..
     generic
       with function "<" ..        --  callee
     package Generic_Sorting is
        function Is_Sorted ...;    --  caller
     end;
  end;

This iteration crashed when the predefined generic unit (which in almost all
cases is a generic package) was inside a generic subprogram (which might
easily happen for user-defined generics). Now fixed.

* src/spark/spark_util.adb
(Enclosing_Generic_Instance): For an instance of a generic subprogram
return its wrapper package, not the instance itself.

* src/spark/spark_util.ads
(Generic_Actual_Subprograms): Stronger precondition that fails on instances
of generic subprograms.

* testsuite/gnatprove/tests/T525-030__flow_generic_actual/test.opt
Enable; no output, just no longer crashes.

Change-Id: I8c312a82e1545b44b1fd7de0f241c9d9f6b72018