T529-035 introduce a temporary for an array formal with relaxed init
On a subprogram call, if the actual and the formal both have compatible
array types but one has relaxed initialization and not the other, we need
a temporary for the call.
Additionally, we were missing a check on out parameters, when the formal
has relaxed init but not the actual. This was due to
Reconstruct_Expr_From_Item not returning an expression of the expected
type and New_Assignment silently introducing a simple conversion.
To avoid this issue in the future, New_Assignment now has a precondition
to ensure that the expression to assign has the expected type. So this
assertion do not fail, we introduce simple conversions to the expected type
before calling New_Assignment in Transform_Assignment_Statement and
Reconstruct_Expr_From_Item once we believe all required checks have been
introduced. We also had to fix Insert_Record_Conversion and
Insert_Pointer_Conversion to change the type of the returned Why node
manually when no conversions are needed.

* src/why/gnat2why-expr.adb
(Get_Item_From_Expr): Update contract so that we do not need to erase the
 init flag of the pattern.
(Get_Item_From_Var): We cannot reuse the actual for an array content if
 the formal and the actual do not have the same relaxed initialization
 status.
(New_Assignment): In New_Assignment, we sometimes introduce an unchecked
 conversion to the expected type of the prefix (when the lvalue is a
 variable). This can result in missing checks on conversions if an
 expression is given with an incorrect type. Add a precondition to check
 that the expression always has the expected type.
(Compute_Call_Args): Do not remove the init flag, we need it for
 Compute_Store.
(Reconstruct_Expr_From_Item): Introduce an initialization check if the
 reconstruction warants one. Convert to the expected type of the actual.
(Transform_Assignment_Statement): Convert to the expected type before
 calling New_Assignment.

* src/why/why-gen-binders.adb
(Item_Array_Length): New boolean parameter to ignore the init flags in the
 length of the item array.

* src/why/why-gen-expr.adb
(Insert_Record_Conversion, Insert_Pointer_Conversion): Make sure that the
 expression has the proper type in gnat2why after the conversion.

* tests/R427-016__check_on_param_call/test.out
* tests/S415-012__list/test.out
Minor changes in output.

* tests/T212-013__relaxed_init_loops/test.out
Change in sloc due to initialization checks now being done as part of the
reconstruction and not as part of the assignment.

* tests/QC11-016__for_of_cntex/test.out
We lose another counterexample for *for of* quantification. This is a
shame but I did not manage to regain it.

* test/SC03-021__binary_search_trees
Update session test.

Change-Id: If8f447322fdc7efabfbb258851e0de890868d79e